chore: remove old layout component
